Revert "Feature/store theme in cookie value (#14)"
This reverts commit 3ba3a5d4881bc9b6f4495be48970c0ed6d8bb557.
