--- src/state/reducers/theme/constants.js
@@ -1,3 +1,0 @@
-const THEME_COOKIE = 'theme';
-
-export default THEME_COOKIE;
