--- src/state/reducers/theme/cookieUtilities/constants.js
@@ -1,2 +1,0 @@
-export const COOKIE_MAX_AGE_TO_ERASE = 'Max-Age=-99999999';
-export const COOKIE_PATH = 'path=/';
