--- src/state/reducers/theme/cookieUtilities/index.js
@@ -1,5 +1,0 @@
-import eraseCookie from './eraseCookie';
-import getCookie from './getCookie';
-import setCookie from './setCookie';
-
-export { eraseCookie, getCookie, setCookie };
