+++ src/state/selectors/index.js
@@ -1,4 +1,3 @@
 export const getTheme = state => state.theme.theme;
-export const getCheckboxCheckedStatus = state => state.theme.checkBoxChecked;
 export const getSections = state => state.sections.sections;
 export const getSectionsContentFetched = state => state.sections.sectionsContentFetched;
