+++ src/state/selectors/index.test.js
@@ -1,6 +1,4 @@
-import {
-  getTheme, getSections, getSectionsContentFetched, getCheckboxCheckedStatus
-} from '.';
+import { getTheme, getSections, getSectionsContentFetched } from '.';
 
 describe('getTheme', () => {
   it('returns the theme property in state', () => {
@@ -29,19 +27,6 @@ describe('getSectionsContentFetched', () => {
       }
     )).toEqual(
       mockSectionsContentFetched
-    );
-  });
-});
-
-describe('getCheckboxCheckedStatus', () => {
-  it('returns the checkBoxChecked property in state', () => {
-    const mockGetCheckboxCheckedStatus = Symbol('test-sections-content-fetched');
-    expect(getCheckboxCheckedStatus(
-      {
-        theme: { checkBoxChecked: mockGetCheckboxCheckedStatus }
-      }
-    )).toEqual(
-      mockGetCheckboxCheckedStatus
     );
   });
 });
